// src/app/modules/master-forms/Models/Company-Devision.ts
export class CompanyDivisionModel {
    BranchId:number = 0;
    CompanyId:number = 0;
    DivisionId: number = 0;
    DivisionCode: string = '';
    DivisionName: string = '';
    FloorName: string = '';
    IsColdStorage: string = '';
    IsActive: string = '';
    AddedBy: string = '';
    Action: string = '';
}